Kp 7367

+++ kcommons/src/main/java/com/kairos/dto/activity/shift/ShiftActivityDTO.java
@@ -53,7 +53,7 @@ public class ShiftActivityDTO {
     private ReasonCodeDTO reasonCode;
     private Long allowedBreakDurationInMinute;
 
-    private int timeBankCtaBonusMinutes;
+    private Double timeBankCtaBonusMinutes;
     private List<TimeBankDistributionDTO> timeBankCTADistributions = new ArrayList<>();
     private Map<String, Object> location;// location where this activity needs to perform
     private String description;// this is from activity description and used in shift detail popup
